skills/cesium-mcp-app/scripts/requirements.txt
mcp>=1.1.0
python-dotenv>=1.0.0
